I made quite a few changes since my last commit.
PlayGUI is updated to print the initial draw and the draw from the player hits. It is tested with ONE player and a dealer. It has an error that appears randomly. I believe it only appears if the hit amount goes past 3 extra cards (that is, 5 total). I'm not sure what the error is exactly, but it doesn't stop the GUI from playing the game. It just prints the error in the console.

GameCoordinator works along with a new class called FirstDrawCollector. FirstDrawCollector has a bit placed right in Deck.java. Every time a card is drawn the ImageIcon is added to FirstDrawCollector. FirstDrawCollector can get the top card (for hits) and will also sort the initial list (called by GameCoordinator in reset()).

FirstDrawCollector is where GameCoordinator gets all the ArrayLists<ImageIcons> and single ImageIcons for the  GUI.